Now tests in separate project also specified os version
