+++ cs/TagsCloudVisualization/SpiralGenerator.cs
@@ -2,9 +2,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace TagsCloudVisualization;
