+++ cs/tdd.sln
@@ -7,7 +7,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BowlingGame", "BowlingGame\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Samples", "Samples\Samples.csproj", "{B5108E20-2ACF-4ED9-84FE-2A718050FC94}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualization", "TagsCloudVisualization\TagsCloudVisualization.csproj", "{80C618CA-794C-4702-99FE-9E65901373D8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TagsCloudVisualization", "TagsCloudVisualization\TagsCloudVisualization.csproj", "{80C618CA-794C-4702-99FE-9E65901373D8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualizationTests", "TagsCloudVisualizationTests\TagsCloudVisualizationTests.csproj", "{F914F9DE-477F-4D5B-AA26-0EDCA99527A6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{80C618CA-794C-4702-99FE-9E65901373D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{80C618CA-794C-4702-99FE-9E65901373D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{80C618CA-794C-4702-99FE-9E65901373D8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F914F9DE-477F-4D5B-AA26-0EDCA99527A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F914F9DE-477F-4D5B-AA26-0EDCA99527A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F914F9DE-477F-4D5B-AA26-0EDCA99527A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F914F9DE-477F-4D5B-AA26-0EDCA99527A6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
